controller update

+++ src/controllers/transactionControllers.js
@@ -4,28 +4,35 @@ import { getUserID } from "./userControllers.js";
 export const getTransactions = async (req, res) => {
   try {
     const username = req?.user?.username;
-
     if (!username) return res.sendStatus(400);
 
     let userID = await getUserID(username);
     if (!userID) return res.sendStatus(401);
 
+    const type = req?.query?.type ?? "latest";
+    const page = +req?.query?.page ?? 1;
+    const ipp = req?.query?.ipp ?? 20;
+
     const startDate = req?.query?.startDate ?? "";
     const endDate = req?.query?.endDate ?? "";
 
-    let data = await Transaction.find({
-      userID: userID,
-      date: {
+    const filters = { userID };
+
+    if (type === "latest") {
+    } else if (type === "period") {
+      filters.date = {
         $gte: new Date(startDate),
         $lte: new Date(endDate),
-      },
-    });
-
-    if (!data) {
-      return res.sendStatus(204);
-    } else {
-      return res.status(200).json(data);
+      };
     }
+
+    const data = await Transaction.find(filters)
+      .limit(ipp)
+      .skip(ipp * (+page - 1));
+
+    const count = await Transaction.countDocuments(filters);
+
+    return res.status(200).json({ data, count });
   } catch (error) {
     console.log(error);
     return res.sendStatus(500);
@@ -39,11 +46,21 @@ export const addTransaction = async (req, res) => {
 
     if (!transaction) return res.sendStatus(400);
 
-    let userID = await getUserID(username);
+    const userID = await getUserID(username);
     if (!userID) return res.sendStatus(401);
 
-    let { id, type, category, description, date, paymethod, amount, currency } =
-      transaction;
+    const {
+      id,
+      type,
+      category,
+      description,
+      details,
+      notes,
+      date,
+      paymethod,
+      amount,
+      currency,
+    } = transaction;
 
     const data = await Transaction.create({
       userID,
@@ -51,15 +68,16 @@ export const addTransaction = async (req, res) => {
       type,
       category,
       description,
+      details,
+      notes,
       date: new Date(date),
       paymethod,
       amount,
       currency,
     });
 
-    return res.status(200).json(data);
+    return res.sendStatus(204);
   } catch (error) {
-    console.log(error);
     return res.sendStatus(500);
   }
 };
@@ -71,13 +89,23 @@ export const editTransaction = async (req, res) => {
 
     if (!transaction) return res.sendStatus(400);
 
-    let userID = await getUserID(username);
+    const userID = await getUserID(username);
     if (!userID) return res.sendStatus(401);
 
-    let { id, type, category, description, date, paymethod, amount, currency } =
-      transaction;
+    const {
+      id,
+      type,
+      category,
+      description,
+      details,
+      notes,
+      date,
+      paymethod,
+      amount,
+      currency,
+    } = transaction;
 
-    let data = await Transaction.updateOne(
+    const data = await Transaction.updateOne(
       {
         userID,
         id,
@@ -87,6 +115,8 @@ export const editTransaction = async (req, res) => {
           type,
           category,
           description,
+          details,
+          notes,
           date,
           paymethod,
           amount,
@@ -95,26 +125,22 @@ export const editTransaction = async (req, res) => {
       }
     );
 
-    return res.status(200).json(data);
+    return res.sendStatus(204);
   } catch (error) {
+    console.log(error);
     return res.sendStatus(500);
   }
 };
 
 export const deleteTransaction = async (req, res) => {
   try {
-    const action = req?.body?.action;
-    const userName = action?.payload?.userName;
+    const id = req?.body?.id;
+    const username = req?.user?.username;
 
-    let userID = await getUserID(userName);
+    let userID = await getUserID(username);
     if (!userID) return res.sendStatus(401);
 
-    const id = action?.payload?.id;
-
-    let data = await Transaction.deleteOne({
-      userID: userID,
-      id: id,
-    });
+    let data = await Transaction.deleteOne({ userID, id });
 
     return res.sendStatus(204);
   } catch (error) {
